Encapsulate "reactive" behaviour in AbstractReactiveStreamsTask class
